Move class dependency analysis to build-update-utils
