+++ build-logic/build-update-utils/src/main/kotlin/gradlebuild/refactoring/tasks/AnalyzeClassDependenciesTask.kt
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package gradlebuild.refactoring
+package gradlebuild.refactoring.tasks
 
 import com.google.gson.GsonBuilder
+import gradlebuild.refactoring.analysis.DependencyAnalyzer
 import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
 import org.gradle.api.artifacts.component.ComponentIdentifier
