--- build-logic/refactoring/build.gradle.kts
@@ -1,11 +1,0 @@
-plugins {
-    id("gradlebuild.build-logic.kotlin-dsl-gradle-plugin")
-}
-
-description = "Provides plugins to help refactor Gradle projects"
-
-dependencies {
-    implementation("gradlebuild:basics")
-    implementation(libs.asm)
-    implementation(buildLibs.gson)
-}
